Alerts and error messages added

+++ src/main/java/app/controller/TextbookController.java
@@ -7,6 +7,7 @@ import app.service.TextbookService;
 import javafx.collections.FXCollections;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
@@ -15,20 +16,20 @@ import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.control.cell.TextFieldTableCell;
 import javafx.scene.input.KeyCode;
-import javafx.scene.layout.AnchorPane;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
-public class TextbookController extends AnchorPane {
+public class TextbookController extends AbstractController {
   private static final Logger logger = LogManager.getLogger(TextbookController.class);
 
-  private final String authorsPattern = "^(\\s*[a-zA-Z]+\\s+[a-zA-Z]+(,\\s+(?!$)|\\s*$))+";
   private final String numberPattern = "^\\s*[0-9]+\\s*$";
 
   private TextbookService textbookService;
@@ -60,9 +61,6 @@ public class TextbookController extends AnchorPane {
         e -> {
           try {
             String title = titleTextField.getText();
-            if (StringUtils.isBlank(title)) {
-              // TODO alert
-            }
             Integer volume = null;
             Integer edition = null;
             String volumeStr = volumeCB.getSelectionModel().getSelectedItem();
@@ -71,14 +69,18 @@ public class TextbookController extends AnchorPane {
               if (volumeStr.matches(numberPattern)) {
                 volume = Integer.valueOf(volumeStr.trim());
               } else {
-                // TODO alert
+                info.setContentText("Volume should be numeric value");
+                info.showAndWait();
+                return;
               }
             }
             if (editionStr != null) {
               if (editionStr.matches(numberPattern)) {
                 edition = Integer.valueOf(editionStr.trim());
               } else {
-                // TODO: alert
+                info.setContentText("Edition should be numeric value");
+                info.showAndWait();
+                return;
               }
             }
             List<Author> authors = authorsFromStr(authorsTextField.getText());
@@ -86,11 +88,16 @@ public class TextbookController extends AnchorPane {
                 FXCollections.observableArrayList(
                     textbookService.searchTextbooks(title, volume, edition, authors)));
           } catch (SQLException e1) {
-            // TODO alert
-            e1.printStackTrace();
+            logger.catching(Level.ERROR, e1);
+            error.setContentText(
+                "Can't perform search due to some internal error.\n" + "See logs for details.");
+            error.showAndWait();
           } catch (InputException e1) {
-            // TODO alert
-            e1.printStackTrace();
+            logger.catching(Level.WARN, e1);
+            info.setContentText(
+                "Authors' first and last names should be specified in sequence, comma separated.\n"
+                    + "For example, Vasya Pupkin, Fedya Nozkin");
+            info.showAndWait();
           }
         });
     addBookButton.setOnAction(
@@ -101,31 +108,36 @@ public class TextbookController extends AnchorPane {
             String editionStr = editionCB.getSelectionModel().getSelectedItem();
             List<Author> authors = authorsFromStr(authorsTextField.getText());
             if (StringUtils.isBlank(title)) {
-              // TODO alert
+              info.setContentText("Book's title shouldn't be blank");
+              info.showAndWait();
+            } else if (!StringUtils.isBlank(volumeStr) && !volumeStr.matches(numberPattern)) {
+              info.setContentText("Book's volume value is entered wrong");
+              info.showAndWait();
+            } else if (StringUtils.isBlank(editionStr) || editionStr.matches(numberPattern)) {
+              info.setContentText("Book's edition isn't specified or isn't numeric value");
+              info.showAndWait();
+            } else if (authors.size() == 0) {
+              info.setContentText("Book should have at least one author");
+              info.showAndWait();
+            } else {
+              Integer volume = Integer.valueOf(volumeStr);
+              Integer edition = Integer.valueOf(editionStr);
+              textbookTableView
+                  .getItems()
+                  .add(textbookService.createTextbook(title, edition, volume, authors));
+              textbookTableView.scrollTo(textbookTableView.getItems().size() - 1);
             }
-            if (!StringUtils.isBlank(volumeStr) && !volumeStr.matches(numberPattern)) {
-              // TODO alert
-            }
-            if (StringUtils.isBlank(editionStr) || editionStr.matches(numberPattern)) {
-              // TODO alert
-            }
-            if (authors.size() == 0) {
-              // TODO alert
-            }
-            Integer volume = Integer.valueOf(volumeStr);
-            Integer edition = Integer.valueOf(editionStr);
-            if (authors.size() == 0) {
-              // TODO alert
-            }
-            textbookTableView
-                .getItems()
-                .add(textbookService.createTextbook(title, edition, volume, authors));
           } catch (SQLException e1) {
-            // TODO alert
-            e1.printStackTrace();
+            logger.catching(Level.ERROR, e1);
+            error.setContentText(
+                "Can't add new algorithm due to some internal error.\n" + "See logs for details.");
+            error.showAndWait();
           } catch (InputException e1) {
-            // TODO alert
-            e1.printStackTrace();
+            logger.catching(Level.WARN, e1);
+            error.setContentText(
+                "Authors' first and last names should be specified in sequence, comma separated.\n"
+                    + "For example, Vasya Pupkin, Fedya Nozhkin");
+            error.showAndWait();
           }
         });
   }
@@ -141,21 +153,7 @@ public class TextbookController extends AnchorPane {
 
     bookIdColumn.setEditable(false);
     titleColumn.setCellFactory(TextFieldTableCell.forTableColumn());
-    volumeColumn.setCellFactory(
-        param ->
-            new TableCell<Textbook, Integer>() {
-              @Override
-              public void updateItem(Integer item, boolean empty) {
-                super.updateItem(item, empty);
-                if (empty || item == null) {
-                  setText(null);
-                } else if (item.equals(0)) {
-                  setText("None");
-                } else {
-                  setText(item.toString());
-                }
-              }
-            });
+    volumeColumn.setCellFactory(param -> new VolumeCell());
     editionColumn.setCellFactory(
         param ->
             new TableCell<Textbook, Integer>() {
@@ -175,13 +173,20 @@ public class TextbookController extends AnchorPane {
           if (e.getCode().equals(KeyCode.DELETE)) {
             Textbook textbook = textbookTableView.getSelectionModel().getSelectedItem();
             if (textbook != null) {
-              // TODO: ask to confirm
-              try {
-                textbookService.deleteTextbook(textbook);
-                textbookTableView.getItems().remove(textbook);
-              } catch (SQLException e1) {
-                // TODO alert
-                e1.printStackTrace();
+              confirm.setContentText(
+                  "Do you really want to delete thid book?\n" + "Operation is undoable.");
+              Optional<ButtonType> buttonType = confirm.showAndWait();
+              if (buttonType.isPresent() && buttonType.get().equals(ButtonType.OK)) {
+                try {
+                  textbookService.deleteTextbook(textbook);
+                  textbookTableView.getItems().remove(textbook);
+                } catch (SQLException e1) {
+                  logger.catching(Level.ERROR, e1);
+                  error.setContentText(
+                      "Can't delete selected book due to some internal error.\n"
+                          + "See logs for details.");
+                  error.showAndWait();
+                }
               }
             }
           }
@@ -190,7 +195,8 @@ public class TextbookController extends AnchorPane {
         e -> {
           String newTitle = e.getNewValue();
           if (StringUtils.isBlank(newTitle)) {
-            // TODO alert
+            info.setContentText("Book's title should be not blank");
+            info.showAndWait();
             textbookTableView.refresh();
           } else {
             try {
@@ -199,33 +205,21 @@ public class TextbookController extends AnchorPane {
               textbookService.updateTitle(textbook, newTitle);
               textbook.setTitle(newTitle);
             } catch (SQLException e1) {
-              // TODO alert
-              e1.printStackTrace();
+              logger.catching(Level.ERROR, e1);
+              error.setContentText(
+                  "Book's title wasn't updated due to some internal error.\n"
+                      + "See logs for details.");
+              error.showAndWait();
             }
           }
         });
 
-    volumeColumn.setOnEditCommit(
-        e -> {
-          Integer newVolume = e.getNewValue();
-          if (newVolume == null || newVolume < 1) {
-            // TODO alert
-            textbookTableView.refresh();
-          } else {
-            try {
-              Textbook textbook = textbookTableView.getItems().get(e.getTablePosition().getRow());
-              textbookService.updateVolume(textbook, newVolume);
-              textbook.setVolume(newVolume);
-            } catch (SQLException e1) {
-              e1.printStackTrace();
-            }
-          }
-        });
     editionColumn.setOnEditCommit(
         e -> {
           Integer newEdition = e.getNewValue();
           if (newEdition == null || newEdition < 1) {
-            // TODO alert
+            info.setContentText("Edition should be specified and should be greater than 0");
+            info.showAndWait();
             textbookTableView.refresh();
           } else {
             try {
@@ -233,8 +227,10 @@ public class TextbookController extends AnchorPane {
               textbookService.updateEdition(textbook, newEdition);
               textbook.setEdition(newEdition);
             } catch (SQLException e1) {
-              // TODO alert
-              e1.printStackTrace();
+              logger.catching(Level.ERROR, e1);
+              error.setContentText(
+                  "Edition wasn't changed due to some internal error.\n" + "See logs for details.");
+              error.showAndWait();
             }
           }
         });
@@ -243,32 +239,40 @@ public class TextbookController extends AnchorPane {
       textbookTableView.setItems(
           FXCollections.observableArrayList(textbookService.getAllTextbooks()));
     } catch (SQLException e) {
-      // TODO alert
-      e.printStackTrace();
+      logger.catching(Level.ERROR, e);
+      info.setContentText(
+          "Can't load books' information due to some internal error.\n" + "See logs for details.");
+      info.showAndWait();
     }
   }
 
   private List<Author> authorsFromStr(String authors) throws InputException {
-    if (authors != null && authors.matches(authorsPattern)) {
-      List<Author> result = new ArrayList<>();
-      if (StringUtils.isBlank(authors)) {
-        return result;
-      }
-      String[] fullNames = authors.split(",\\s+");
-      for (String fullName : fullNames) {
-        String[] fullNameArr = fullName.split("\\s+");
-        result.add(new Author(-1, fullNameArr[0], fullNameArr[1]));
-      }
-      return result;
+    if (StringUtils.isBlank(authors)) {
+      return new ArrayList<>();
     } else {
-      throw new InputException("Authors' list doesn't match the pattern");
+      String authorsPattern = "^(\\s*[a-zA-Z]+\\s+[a-zA-Z]+(,\\s+(?!$)|\\s*$))+";
+      if (authors.matches(authorsPattern)) {
+        List<Author> result = new ArrayList<>();
+        if (StringUtils.isBlank(authors)) {
+          return result;
+        }
+        String[] fullNames = authors.split(",\\s+");
+        for (String fullName : fullNames) {
+          String[] fullNameArr = fullName.split("\\s+");
+          result.add(new Author(-1, fullNameArr[0], fullNameArr[1]));
+        }
+        return result;
+      } else {
+        throw logger.throwing(
+            Level.WARN, new InputException("Authors' list doesn't match the pattern"));
+      }
     }
   }
 
   private class AuthorCell extends TableCell<Textbook, List<Author>> {
     private TextField textField;
 
-    public AuthorCell() {
+    AuthorCell() {
       textField = new TextField();
       textField
           .focusedProperty()
@@ -285,9 +289,12 @@ public class TextbookController extends AnchorPane {
                 List<Author> authorsFromInput = authorsFromStr(textField.getText());
                 commitEdit(authorsFromInput);
               } catch (InputException e1) {
-                // TODO alert
+                logger.catching(Level.WARN, e1);
+                info.setContentText(
+                    "Authors' first and last names should be specified in sequence, comma separated.\n"
+                        + "For example, Vasya Pupkin, Fedya Nozkin");
+                info.showAndWait();
                 cancelEdit();
-                e1.printStackTrace();
               }
             }
           });
@@ -325,9 +332,11 @@ public class TextbookController extends AnchorPane {
         setItem(newAuthors);
         updateItem(newAuthors, false);
       } catch (SQLException e) {
-        // TODO alert
-        e.printStackTrace();
-        cancelEdit();
+        logger.catching(Level.ERROR, e);
+        error.setContentText(
+            "Authors weren't updated due to some internal error.\n" + "See logs for details.");
+        error.showAndWait();
+        textbookTableView.refresh();
       }
     }
 
@@ -352,6 +361,110 @@ public class TextbookController extends AnchorPane {
         setGraphic(null);
         setText(getString());
       }
+    }
+  }
+
+  private class VolumeCell extends TableCell<Textbook, Integer> {
+    private TextField textField;
+
+    VolumeCell() {
+      textField = new TextField();
+      textField
+          .focusedProperty()
+          .addListener(
+              (obs, oldValue, newValue) -> {
+                if (newValue && !oldValue) {
+                  textField.selectAll();
+                }
+              });
+      textField.setOnKeyReleased(
+          e -> {
+            if (e.getCode().equals(KeyCode.ENTER)) {
+              String volumeStr = textField.getText();
+              if (StringUtils.isBlank(volumeStr)) {
+                setItem(null);
+                try {
+                  Textbook rowValue = textbookTableView.getItems().get(getIndex());
+                  textbookService.updateVolume(rowValue, null);
+                  rowValue.setVolume(null);
+                } catch (SQLException e1) {
+                  logger.catching(Level.ERROR, e1);
+                  error.setContentText(
+                      "Can't set book's volume to none due to some internal error.\n"
+                          + "See logs for details.");
+                  error.showAndWait();
+                }
+              } else if (volumeStr.matches(numberPattern)) {
+                commitEdit(Integer.valueOf(volumeStr.trim()));
+              } else {
+                info.setContentText("Volume should be numeric value");
+                info.showAndWait();
+                cancelEdit();
+              }
+            }
+          });
+    }
+
+    @Override
+    protected void updateItem(Integer item, boolean empty) {
+      super.updateItem(item, empty);
+      if (empty) {
+        setText(null);
+        setGraphic(null);
+      } else {
+        if (isEditing()) {
+          textField.setText(getString());
+          setText(null);
+          setGraphic(textField);
+        } else {
+          setGraphic(null);
+          setText(getString());
+        }
+      }
+    }
+
+    @Override
+    public void startEdit() {
+      if (!isEmpty()) {
+        super.startEdit();
+        setText(null);
+        textField.setText(getText());
+        setGraphic(textField);
+        textField.requestFocus();
+      }
+    }
+
+    @Override
+    public void commitEdit(Integer newValue) {
+      super.commitEdit(newValue);
+      if (newValue < 1) {
+        info.setContentText("Book's volume should be greater than 0");
+        info.showAndWait();
+        cancelEdit();
+      } else {
+        try {
+          Textbook rowValue = textbookTableView.getItems().get(getIndex());
+          textbookService.updateVolume(rowValue, newValue);
+          rowValue.setVolume(newValue);
+        } catch (SQLException e) {
+          logger.catching(Level.ERROR, e);
+          error.setContentText(
+              "Can't set book's volume to none due to some internal error.\n"
+                  + "See logs for details.");
+          error.showAndWait();
+        }
+      }
+    }
+
+    @Override
+    public void cancelEdit() {
+      super.cancelEdit();
+      setGraphic(null);
+      setText(getString());
+    }
+
+    private String getString() {
+      return getItem() == null ? "None" : getItem().toString();
     }
   }
 }
